+++ src/database-metrics-logger.ts
@@ -10,7 +10,7 @@ import {
 } from './interfaces/database-metrics-logger-config.interface';
 import { MongoDbAgent } from './modules/database-metrics/mongodb/agent';
 import { RedisAgent } from './modules/database-metrics/redis/agent';
-import { ITransportInterface } from './modules/transports/transport-interface';
+import { ITransportInterface } from './modules/transports/interfaces/transport-interface';
 
 const defaultOptions = {
   interval: 10000,
@@ -69,7 +69,7 @@ export class DatabaseMetricsLogger extends PubSub {
 
   private executeTransports(metrics: IMetricsResponse): void {
     if (this.transports) {
-      this.transports.forEach(transport => transport.postMetrics(metrics));
+      this.transports.forEach(transport => transport.post(metrics));
     }
   }
 }