+++ src/modules/transports/datadog/database-definitions/mongodb-definition.ts
@@ -1,10 +1,11 @@
+import { LogStatus } from '../enums';
 import { IDatabaseDefinition } from '../interfaces/database-definition.interface';
 
 export const mongoDbDefinition: IDatabaseDefinition = {
   metricMaps: {
     // read requests per second
-    'mongodb.opcounters.getmoreps': 'serverStatus.opcounters.getmore',
-    'mongodb.opcounters.queryps': 'serverStatus.opcounters.query',
+    'mongodb.opcounters.getmoreps': 'metrics.serverStatus.opcounters.getmore',
+    'mongodb.opcounters.queryps': 'metrics.serverStatus.opcounters.query',
 
     // write requests per second
     'mongodb.opcounters.deleteps': 'serverStatus.opcounters.delete',
@@ -72,10 +73,15 @@ export const mongoDbDefinition: IDatabaseDefinition = {
     'mongodb.replset.secondarymemberscount': 'custom.replicationSetSecondaryMembersCount',
     'mongodb.replset.activememberscount': 'custom.replicationSetActiveMembersCount',
     'mongodb.replset.inactivememberscount': 'custom.replicationSetInactiveMembersCount',
-
-    // version (custom)
-    'mongodb.version': 'serverStatus.version',
   },
+  logs: [{
+    message: 'MongoDB Version',
+    status: LogStatus.Info,
+    ddsource: 'mongodb',
+    attributesMap: {
+      version: 'serverStatus.version',
+    },
+  }],
   tagMaps: {
     'replset_name': 'custom.replicationSetName',
     'mongodb-version': 'serverStatus.version',