+++ src/modules/transports/datadog/enums/default-hosts.enum.ts
@@ -1,0 +1,4 @@
+export enum DefaultHost {
+  Host = 'http-intake.logs.datadoghq.com',
+  LogHost = 'http-intake.logs.datadoghq.com',
+}