+++ src/modules/transports/datadog/enums/event-priority.enum.ts
@@ -1,0 +1,4 @@
+export enum EventPriority {
+  Normal = 'normal',
+  Low = 'low',
+}