+++ src/modules/transports/datadog/interfaces/datadog-options.ts
@@ -2,5 +2,6 @@ export interface IDatadogOptions {
   apiKey: string;
   appKey?: string;
   host?: string;
+  logHost?: string;
   tags?: string[];
 }