+++ src/modules/transports/datadog/interfaces/event-response.interface.ts
@@ -1,0 +1,7 @@
+import { ResponseStatus } from '../enums';
+import { IEventResponseEvent } from './event-response-event.interface';
+
+export interface IEventResponse {
+  event: IEventResponseEvent;
+  status: ResponseStatus;
+}