+++ src/modules/transports/datadog/interfaces/log-request.interface.ts
@@ -1,0 +1,7 @@
+export interface ILogRequest {
+  message?: string;
+  ddsource?: string;
+  hostname?: string;
+  ddtags?: string;
+  [key: string]: string;
+}