+++ src/modules/transports/interfaces/transport-interface.ts
@@ -1,0 +1,5 @@
+import { IMetricsResponse } from '../../../interfaces';
+
+export interface ITransportInterface {
+  post(metrics: IMetricsResponse): Promise<any>;
+}