+++ src/modules/transports/transport-interface.ts
@@ -1,5 +1,0 @@
-import { IMetricsResponse } from '../../interfaces';
-
-export interface ITransportInterface {
-  postMetrics(metrics: IMetricsResponse): Promise<any>;
-}